+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Implemented enhancements
 
+- GP2-1571 - Case study - Wagtail quote block
 - GP2-1619 - Ensure all pages ported from V1 populate GA360/DataLayer data
 - GP2-1180 - Travel and business policies-page
 - GP2-1404 - Select your product (modal window) - Product and country are not selected